[36-FEAT] 약속 제안 확정 API 구현

+++ models/user.ts
@@ -17,7 +17,7 @@ import EventModel from './event';
 @Table({ tableName: 'User', modelName: 'User' })
 class User extends Model {
   @PrimaryKey
-  @Column({ type: DataType.INTEGER, field: 'userId' })
+  @Column({ type: DataType.BIGINT, field: 'userId' })
   id: number;
 
   @AllowNull(false)
